Move Triangles to Shapes
